# openapi.yaml
openapi: 3.0.3
info:
  title: Point of Sale (POS) System API
  version: 1.0.0
  description: >
    Comprehensive REST API documentation for POS System Backend featuring multi-branch management,
    real-time analytics, inventory control, cashier shift management, refund processing, and payment integration.
  contact:
    name: POS System Support
    email: support@pos-system.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
security:
  - BearerAuth: []
paths:
...
  /api/store/{id}:
    get:
      tags:
        - Store Management
      summary: Get Store by ID
      description: Retrieves store details by store ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved store details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoreDTO'
    put:
      tags:
        - Store Management
      summary: Update Store
      description: Updates an existing store's profile by store ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreDTO'
      responses:
        '200':
          description: Store successfully updated
    delete:
      tags:
        - Store Management
      summary: Delete Store
      description: Deletes a store by store ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Store successfully deleted
...
  /api/branch/{id}:
    get:
      tags:
        - Branch Management
      summary: Get Branch by ID
      description: Retrieves branch details by branch ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved branch details
    put:
      tags:
        - Branch Management
      summary: Update Branch
      description: Updates an existing branch's details by branch ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchDTO'
      responses:
        '200':
          description: Branch successfully updated
    delete:
      tags:
        - Branch Management
      summary: Delete Branch
      description: Deletes a branch by branch ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Branch successfully deleted
...
  /api/employee/store/{storeId}:
    post:
      tags:
        - Employee Management
      summary: Create Store Employee
      description: Creates a new employee (e.g. Store Admin/Manager) associated with a store.
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '201':
          description: Store employee successfully created
    get:
      tags:
        - Employee Management
      summary: Find All Store Employees
      description: Retrieves all employees for a store, optionally filtered by user role.
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userRole
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved store employees list
  /api/employee/branch/{branchId}:
    post:
      tags:
        - Employee Management
      summary: Create Branch Employee
      description: Creates a new employee (e.g. Cashier, Branch Manager) assigned to a specific branch.
      parameters:
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '201':
          description: Branch employee successfully created
    get:
      tags:
        - Employee Management
      summary: Find All Branch Employees
      description: Retrieves all employees assigned to a specific branch, optionally filtered by role.
      parameters:
        - name: branchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userRole
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved branch employees list
  /api/employee/{employeeId}:
    put:
      tags:
        - Employee Management
      summary: Update Employee
      description: Updates employee details (name, email, role, phone) by employee ID.
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
      responses:
        '200':
          description: Employee successfully updated
    delete:
      tags:
        - Employee Management
      summary: Delete Employee
      description: Removes an employee by employee ID.
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Employee successfully deleted

  # ---------------------------------------------------------------------------
  # Category Management
  # ---------------------------------------------------------------------------
...
  /api/customers:
    post:
      tags:
        - Customer Management
      summary: Create Customer
      description: Registers a new customer in the POS system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerCreateDTO'
      responses:
        '201':
          description: Customer successfully created
    get:
      tags:
        - Customer Management
      summary: Get All Customers
      description: Retrieves a complete list of all registered customers.
      responses:
        '200':
          description: Successfully retrieved customers list
  /api/customers/{id}:
    get:
      tags:
        - Customer Management
      summary: Get Customer by ID
      description: Retrieves customer profile and loyalty details by customer ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved customer details
    put:
      tags:
        - Customer Management
      summary: Update Customer
      description: Updates an existing customer's contact details or loyalty info.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerUpdateDTO'
      responses:
        '200':
          description: Customer successfully updated
    delete:
      tags:
        - Customer Management
      summary: Delete Customer
      description: Deletes a customer by ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Customer successfully deleted
  /api/customers/search:
    get:
      tags:
        - Customer Management
      summary: Search Customers
      description: Searches for customers by name, phone, or email with pagination.
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved paginated search results

  # ---------------------------------------------------------------------------
  # Order Management
  # ---------------------------------------------------------------------------
...
  /api/orders/customer/{customerId}:
    get:
      tags:
        - Order Management
      summary: Get Customer Orders
      description: Retrieves purchase history for a specific customer.
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successfully retrieved customer order history
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter your JWT Bearer token obtained from `/auth/login` or `/auth/signup`
  schemas:
    UserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        password:
          type: string
        fullName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [ROLE_ADMIN, ROLE_STORE_ADMIN, ROLE_STORE_MANAGER, ROLE_BRANCH_MANAGER, ROLE_BRANCH_CASHIER]
        storeId:
          type: integer
          format: int64
        branchId:
          type: integer
          format: int64
...
    StoreDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        brand:
          type: string
        description:
          type: string
        storeAdminId:
          type: integer
          format: int64
        status:
          type: string
          enum: [ACTIVE, INACTIVE, SUSPENDED]
    BranchDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        storeId:
          type: integer
          format: int64
...
    CustomerCreateDTO:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
    CustomerUpdateDTO:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        loyaltyPoints:
          type: integer
...